# apps/api/openapi/formex-api.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Formex",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/programs/{program}/timetable": {
            "get": {
                "operationId": "timetable.show",
                "tags": [
                    "Timetable"
                ],
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "required": true,
                        "description": "The program ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "semester",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "string"
                                                        },
                                                        "semester": {
                                                            "type": "string"
                                                        },
                                                        "academic_year": {
                                                            "type": "string"
                                                        },
                                                        "slots": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string"
                                                                    },
                                                                    "day": {
                                                                        "type": "string"
                                                                    },
                                                                    "start_hour_no": {
                                                                        "type": "integer"
                                                                    },
                                                                    "end_hour_no": {
                                                                        "type": "integer"
                                                                    },
                                                                    "subjects": {
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "required": [
                                                                    "id",
                                                                    "day",
                                                                    "start_hour_no",
                                                                    "end_hour_no",
                                                                    "subjects"
                                                                ]
                                                            }
                                                        }
                                                    },
                                                    "required": [
                                                        "id",
                                                        "semester",
                                                        "academic_year",
                                                        "slots"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "semester": {
                                                            "type": "integer"
                                                        },
                                                        "academic_year": {
                                                            "type": "integer"
                                                        },
                                                        "slots": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            },
                                                            "minItems": 0,
                                                            "maxItems": 0,
                                                            "additionalItems": false
                                                        }
                                                    },
                                                    "required": [
                                                        "semester",
                                                        "academic_year",
                                                        "slots"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "data"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "timetable.upsertSlot",
                "tags": [
                    "Timetable"
                ],
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "required": true,
                        "description": "The program ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "semester": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "maximum": 12
                                    },
                                    "day": {
                                        "type": "string"
                                    },
                                    "start_hour_no": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "maximum": 7
                                    },
                                    "end_hour_no": {
                                        "type": "integer",
                                        "minimum": 2,
                                        "maximum": 7
                                    },
                                    "subjects": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "subject_id": {
                                                    "type": "string"
                                                },
                                                "course_coordinator_id": {
                                                    "type": "string"
                                                },
                                                "batch": {
                                                    "type": "string",
                                                    "maxLength": 255
                                                },
                                                "room_no": {
                                                    "type": "string",
                                                    "maxLength": 255
                                                }
                                            },
                                            "required": [
                                                "subject_id",
                                                "course_coordinator_id",
                                                "batch",
                                                "room_no"
                                            ]
                                        },
                                        "minItems": 1,
                                        "maxItems": 10
                                    }
                                },
                                "required": [
                                    "semester",
                                    "day",
                                    "start_hour_no",
                                    "end_hour_no",
                                    "subjects"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Timetable slot saved successfully."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "day": {
                                                    "type": "string"
                                                },
                                                "start_hour_no": {
                                                    "type": "integer"
                                                },
                                                "end_hour_no": {
                                                    "type": "integer"
                                                },
                                                "subjects": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "day",
                                                "start_hour_no",
                                                "end_hour_no",
                                                "subjects"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
...
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}